Moved User and UserService to service package
