+++ src/main/java/com/company/service/UserServiceImpl.java
@@ -1,6 +1,6 @@
-package main.java.com.company.dao.users.impl;
+package main.java.com.company.service;
 
-import main.java.com.company.dao.users.UserService;
+import main.java.com.company.service.interfaces.UserService;
 import main.java.com.company.model.user.User;
 import main.java.com.company.model.user.enums.Rights;
 import main.java.com.company.model.user.enums.Status;
