+++ src/main/java/com/company/service/interfaces/UserService.java
@@ -1,4 +1,4 @@
-package main.java.com.company.dao.users;
+package main.java.com.company.service.interfaces;
 
 public interface UserService {
     void userAdd(String login, String password);
